% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mariokart.R
\docType{package}
\name{mariokart}
\alias{mariokart}
\alias{mariokart-package}
\title{Mario Kart Statistics from Different Games}
\description{
This package contains complete statistics for Mario Kart games
}
\examples{
data(mk8_characters)

plot(data = mk8_characters, speed_normal ~ handling_normal)
}
