% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk8.R
\docType{data}
\name{mk8_characters}
\alias{mk8_characters}
\title{"Mario Kart 8: Characters"}
\format{A data frame with 16 observations of 14 variables}
\usage{
mk8_characters
}
\description{
A data set containing character statistics for Mario Kart 8.
}
\keyword{datasets}
