% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mk8_gliders}
\alias{mk8_gliders}
\title{"Mario Kart 8: Gliders"}
\format{A data frame with 14 observations of 13 variables}
\usage{
mk8_gliders
}
\description{
A data set containing glider statistics for Mario Kart 8.
}
\keyword{datasets}
