% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mk8_vehicles}
\alias{mk8_vehicles}
\title{"Mario Kart 8: Vehicles"}
\format{A data frame with 40 observations of 15 variables}
\usage{
mk8_vehicles
}
\description{
A data set containing vehicle statistics for Mario Kart 8.
}
\keyword{datasets}
