% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mk8_wheels}
\alias{mk8_wheels}
\title{"Mario Kart 8: Wheels"}
\format{A data frame with 21 observations of 13 variables}
\usage{
mk8_wheels
}
\description{
A data set containing wheels statistics for Mario Kart 8.
}
\keyword{datasets}
