% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mkwii_characters}
\alias{mkwii_characters}
\title{"Mario Kart Wii: Characters"}
\format{A data frame with 26 observations of 8 variables}
\usage{
mkwii_characters
}
\description{
A data set containing character statistics for Mario Kart Wii.
}
\keyword{datasets}
