% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mkwii_vehicles}
\alias{mkwii_vehicles}
\title{"Mario Kart Wii: Vehicles"}
\format{A data frame with 36 observations of 12 variables}
\usage{
mkwii_vehicles
}
\description{
A data set containing vehicle statistics for Mario Kart Wii.
}
\keyword{datasets}
