% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smk_acceleration}
\alias{smk_acceleration}
\title{"Super Mario Kart: Acceleration"}
\format{A data frame with 106 observations of 3 variables}
\usage{
smk_acceleration
}
\description{
A data set containing experimental data for speed with respect to time for Super Mario Kart.
}
\keyword{datasets}
