% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smk_characters}
\alias{smk_characters}
\title{"Super Mario Kart: Character"}
\format{A data frame with 8 observations of 5 variables}
\usage{
smk_characters
}
\description{
A data set containing character statistics for Super Mario Kart.
}
\keyword{datasets}
