% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smk_speeds}
\alias{smk_speeds}
\title{"Super Mario Kart: Speeds"}
\format{A data frame with 8 observations of 8 variables}
\usage{
smk_speeds
}
\description{
A data set containing speed statistics on different terrains for Super Mario Kart.
}
\keyword{datasets}
